refactor: MainApplication 경로 변경
@ComponentScan 을 활용하여 수동으로 주입하는 코드 삭제
