+++ winwin-be-api/src/main/java/com/dpm/winwin/WinwinApiApplication.java
@@ -1,0 +1,12 @@
+package com.dpm.winwin;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WinwinApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(WinwinApiApplication.class);
+    }
+}
