+++ winwin-be-chatting/src/main/java/com/dpm/winwin/WinwinChattingApplication.java
@@ -1,0 +1,13 @@
+package com.dpm.winwin;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class WinwinChattingApplication {
+
+  public static void main(String[] args) {
+      SpringApplication.run(WinwinChattingApplication.class);
+  }
+}
